fix: error reason fn & linting
